res fixed a couple VC7 64bit warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20620 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/demo/demo.cpp
@@ -139,7 +139,7 @@ static void TestDemoFile (const char* zip, iVFS* myVFS, csStringArray& demos)
   if (testpath[strlen (testpath)-1] == '/')
   {
     // We have a directory.
-    int l = strlen (testpath);
+    size_t l = strlen (testpath);
     testpath[l-1] = '$';
     testpath[l] = '/';
     testpath[l+1] = 0;
