+++ apps/import/map2cs/texman.h
@@ -57,10 +57,10 @@ public:
   bool AddAllTexturesToVFS(csRef<iVFS> VFS, const char* path);
 
   /// returns the Number of known textures
-  int GetTextureCount() {return m_StoredTextures.Length();}
+  size_t GetTextureCount() { return m_StoredTextures.Length(); }
 
   /// returns the n-th texture (from 0 to GetTextureCount()-1)
-  CTextureFile* GetTexture(int index) {return m_StoredTextures[index];}
+  CTextureFile* GetTexture (size_t index) {return m_StoredTextures[index];}
 
 protected:
   void CleanupTexturename(char* Name);
